Merge MWE retrieval via lemmata into call via id to /mwe/profile
The previously dysfunctioning endpoint /list/mwe is removed.
Calls to /mwe/profile are now possible with a pair of lemmata (lemma1, lemma2) as well as with coocc_id as before.
If both coocc_id and a pair of lemmata is passed, the id take priority and the strings are ignored.