Change update method for MWE frequencies: ignore hapax legomena
The default value for the dictionary that stores the frequencies of MWE collocations is set to 1 instead of 0. Thus, each MWE collocations is added to the frequency count only if it is encountered a second time. This should somewhat reduce the memory footprint of the MWE computation, since MWE collocations that occur only once are not stored.

The nested function to index the MWE collocations was moved to top-level to make it testable.