Filter MWE id map to remove all entries with frequency of 1
During logDice computation, the MWE collocations are filtered for the desired minimal frequency. So, removing all collocations that don't appear in the id-frequency map (i.e. MWE collocations that were found but not encountered a second time and thus not added to the id-freqency map) will reduce processing time and memory. However, this will also introduce some error as the 'token frequencies' of the collocations will be lower than expected. In a future update, this should be changed to use the frequencies from the collocs dictionary instead of calculating the frequencies from the MWE occurances.