Filter MWE matches to keep only matches for collocations that reach minimal frequency
As before, all MWE matches are written to a common file. After extracting collocations and computing scores, this temporary file containing the matches is processed again and filtered against the MWE collocations. Matches for valid collocations are written to the final file, all others are discarded. Finally, the temporary file is removed.