Use individual counter for each process to count lemma frequencies
Instead of a shared dictionary to store lemma frequencies, each process now creates its own Counter that is passed via a shared list to the main process where the lemma counts are aggregated. As the last step of the data processing, the collected lemma frequencies are written to a file.