feat!: change extraction of PRED collocations (#27)
* feat: add 'werden' und 'bleiben' as valid copula verbs for subject predicative

* test: use local cached DWDSmor files

* feat!: remove extraction of predicative collocations without copula verb

The distinction between predicative collocations with verbs as head and collocations of prepositional phrases is not clear enough. Often prepositional phrases are categorized as PREP collocations, effectivly appearing twice. Detection of PRED collocations with verb as head needs to be revised. Until then, only PRED collocations with copula are extracted.